# server/swagger.yaml
openapi: 3.0.3
info:
  title: Milk Dairy Management System API
  description: |
    A comprehensive MERN stack application for managing milk dairy operations with role-based access control, 
    subscription management, and real-time delivery tracking.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## User Roles
    - **ADMIN**: Full system access, user management, analytics
    - **DISTRIBUTOR**: Order management, delivery assignment, product management
    - **DELIVERY_BOY**: Delivery assignments, status updates
    - **CONSUMER**: Place orders, manage subscriptions, track deliveries
  version: 1.0.0
  contact:
    name: API Support
    email: support@milkdairy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.milkdairy.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /subscriptions:
    get:
      tags:
        - Subscription Management
      summary: Get user subscriptions
      description: Retrieve all subscriptions for the authenticated user
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Subscription Management
      summary: Create new subscription
      description: Create a new subscription for recurring orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
            example:
              products:
                - productId: "60d5ecb74b24a1234567890a"
                  quantity: 2
              subscriptionType: "weekly"
              frequency: "weekly"
              startDate: "2024-01-01"
              endDate: "2024-12-31"
              deliveryTime: "morning"
              deliveryDays: ["monday", "wednesday", "friday"]
              specialInstructions: "Leave at the door"
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subscriptions/{id}:
    get:
      tags:
        - Subscription Management
      summary: Get subscription by ID
      description: Retrieve a specific subscription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: Subscription retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Subscription Management
      summary: Update subscription
      description: Update an existing subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Subscription Management
      summary: Cancel subscription
      description: Cancel an existing subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Subscription already cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /subscriptions/{id}/resume:
    put:
      tags:
        - Subscription Management
      summary: Resume paused subscription
      description: Resume a paused subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      responses:
        '200':
          description: Subscription resumed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found or not in paused state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Admin Endpoints
...
  /consumer/subscriptions:
    get:
      tags:
        - Consumer
      summary: Get consumer subscriptions
      description: Retrieve all subscriptions for the authenticated consumer
      responses:
        '200':
          description: Subscriptions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Consumer
      summary: Create consumer subscription
      description: Create a new subscription for recurring orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerSubscriptionRequest'
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error or no distributor assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /consumer/subscriptions/{subscriptionId}:
    put:
      tags:
        - Consumer
      summary: Update consumer subscription
      description: Update an existing subscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
          description: Subscription ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '403':
          description: Not authorized to update this subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subscription:
      type: object
      properties:
        _id:
          type: string
          description: Subscription ID
        customer:
          type: string
          description: Customer user ID
        distributor:
          type: string
          description: Distributor user ID
        orders:
          type: array
          items:
            type: string
          description: Array of subscription order IDs
        subscriptionType:
          type: string
          enum: [one-time, weekly, monthly, custom]
        products:
          type: array
          items:
            type: object
            properties:
              product:
                type: string
                description: Product ID
              quantity:
                type: integer
                minimum: 1
              price:
                type: number
              unit:
                type: string
        frequency:
          type: string
          enum: [daily, twice_week, alternate_days, weekly, fortnightly, monthly]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        deliveryTime:
          type: string
          enum: [morning, afternoon, evening, flexible]
        deliveryDays:
          type: array
          items:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        deliveryDay:
          type: integer
          minimum: 1
          maximum: 28
        specialInstructions:
          type: string
        status:
          type: string
          enum: [active, paused, cancelled, completed]
        nextDeliveryDate:
          type: string
          format: date-time
        lastDeliveryDate:
          type: string
          format: date-time
        totalDeliveries:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubscriptionRequest:
      type: object
      required:
        - products
        - subscriptionType
        - frequency
        - startDate
      properties:
        products:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
                minimum: 1
          minItems: 1
        subscriptionType:
          type: string
          enum: [one-time, weekly, monthly, custom]
        frequency:
          type: string
          enum: [daily, twice_week, alternate_days, weekly, fortnightly, monthly]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        deliveryTime:
          type: string
          enum: [morning, afternoon, evening, flexible]
          default: morning
        deliveryDays:
          type: array
          items:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        deliveryDay:
          type: integer
          minimum: 1
          maximum: 28
        specialInstructions:
          type: string
    ConsumerSubscriptionRequest:
      type: object
      required:
        - products
        - frequency
        - startDate
        - endDate
      properties:
        products:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
                minimum: 1
          minItems: 1
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        deliveryTime:
          type: string
          enum: [morning, afternoon, evening]
          default: morning
        deliveryAddress:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            pincode:
              type: string
            landmark:
              type: string
        specialInstructions:
          type: string
        deliveryDays:
          type: array
          items:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]

    # Delivery Schema
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (development only)
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Validation errors
...